# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceKernel API",
    "version": "1.0.0",
    "description": "The voice infrastructure layer for regulated enterprise. Create agents, place calls, ground answers in your knowledge, and stream every event into your own systems."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/calls": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "List calls",
        "responses": {
          "200": {
            "description": "A list of calls."
          }
        }
      },
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Place a call",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "to": "+61400000000",
                "agentId": "asst_123",
                "phoneNumberId": "pn_456",
                "metadata": {
                  "crmTicket": "INC-9912"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created call."
          }
        }
      }
    },
    "/v1/calls/{id}": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "Retrieve a call",
        "responses": {
          "200": {
            "description": "The call."
          }
        }
      },
      "delete": {
        "tags": [
          "Calls"
        ],
        "summary": "End a call or delete its data",
        "responses": {
          "200": {
            "description": "Deleted."
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v1/calls/{id}/transcript": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "Transcript, summary and analysis",
        "responses": {
          "200": {
            "description": "The transcript."
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v1/calls/{id}/artifacts/{artifact}": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "Presigned URL for a recording, pcap or call log",
        "responses": {
          "200": {
            "description": "{ url, expiresIn, artifact }"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "artifact",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "recording",
              "mono-recording",
              "stereo-recording",
              "video-recording",
              "customer-recording",
              "assistant-recording",
              "pcap",
              "logs"
            ]
          }
        }
      ]
    },
...
    "/v1/files/{id}": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "Retrieve a file",
        "responses": {
          "200": {
            "description": "The object."
          }
        }
      },
      "patch": {
        "tags": [
          "file"
        ],
        "summary": "Update a file",
        "responses": {
          "200": {
            "description": "Updated."
          }
        }
      },
      "delete": {
        "tags": [
          "file"
        ],
        "summary": "Delete a file",
        "responses": {
          "200": {
            "description": "Deleted."
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "vk_live_…"
      }
    },
...
  }
}